pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "retri-evals"
version = "0.0.1"
homepage = "https://github.com/DeployQL/retri-evals"
description = "Open-source tool for building and evaluating retrieval pipelines."
readme = "README.md"
license = "AGPL-3.0"
keywords = [
    "data-science",
    "artificial-intelligence",
    "developers-tools",
    "mlops",
    "rag",
    "retrieval",
]
authors = [
    "Matt Barta <contact@deployql.com>"
]
packages = [
    {include = "retri_eval"},
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pydantic-settings = "2.1.0"
tantivy = "0.21.0"
qdrant-client = "1.7.0"
tokenizers = "^0.15.0"
beir = "2.0.0"
mteb = "1.1.1"
hruid = "0.0.3"
dspy-ai = "^2.1.1"
ir-datasets = "^0.5.5"
ipython = "^8.20.0"
llama-cpp-python = "^0.2.31"
autoawq = "^0.1.8"
transformers = "^4.36.2"
sentence-transformers = "^2.2.2"
usearch = "^2.8.15"
openai = "<=0.28.1"

[tool.poetry.group.examples]
optional = true

[tool.poetry.group.examples.dependencies]


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"


[tool.poetry.group.dev-dependencies.dependencies]
black = "^23.12.1"

[[tool.poetry.source]]
name = "PyPI"
priority = "default"

